move BarChart component into components folder
